Create and implement collision detection system
Currently, there is an undesired behaviour where after the user collides with a wall in a certain direction, they cannot move perpendicular to that direction until moving the opposite direction to 'uncollide' with the wall. To fix this, we will probably want to push the player away from the wall they are currently colliding with, or consider something else.