Refactored Menu Class and menu components
Created multiple java classes for the various components displayed in the menu, including buttons, backgrounds, banners and labels.

I also created a menuimageloader class to separate the menu images from the image loader and removed the need for enum maps and using enum classes for referencing images as that limits extendability of code.